Move all API calls to api folder and move types to types folder
