+++ frontend/pages/settings.tsx
@@ -1,7 +1,6 @@
 import { Heading, Box, Text, Button, Divider } from "@chakra-ui/react";
 import ThemeToggler from "@components/atoms/ThemeToggler";
 import Header from "@components/molecules/Header";
-import NavBar from "@components/molecules/NavBar";
 import Layout from "@components/templates/Layout";
 import { useTranslation } from "next-i18next";
 import React, { FC } from "react";
