--- frontend/pages/song/index.tsx
@@ -1,7 +1,0 @@
-import React, { FC } from "react";
-
-const Song: FC = () => {
-  return <p>Hei</p>;
-};
-
-export default Song;
