+++ frontend/src/components/molecules/VerseNumbering.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from "react";
 import { Stack, Text } from "@chakra-ui/react";
-import { SongDetailType } from "src/api/songs";
+import { SongDetailType } from "src/types";
 
 interface VerseNumberingProps {
   song: SongDetailType;
