+++ frontend/src/types.ts
@@ -1,0 +1,26 @@
+export type Slug = {
+  current: string;
+};
+
+export type Info = {
+  prev: string | null;
+  next: string | null;
+};
+
+export interface SongListEntry {
+  title: string;
+  slug: Slug;
+}
+
+export interface SongDetailType {
+  title: string;
+  author: string;
+  melody: string;
+  order: number;
+  slug: Slug;
+  verses: string[];
+  category: string;
+  info: Info;
+  chorus: string;
+  spotifyuri: string;
+}
